requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
asyncpg==0.29.0
psycopg==3.1.15  # For sync operations

# Security & Authentication
python-keycloak==3.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography>=41.0.0,<46.0.0

# HTTP Client for service calls
httpx==0.25.2
aiohttp==3.9.1

# Event-driven architecture
celery==5.3.4
redis==5.0.1
kombu==5.3.4

# Background tasks and async
asyncio-mqtt==0.16.1
aiokafka==0.9.0

# System monitoring and debugging
psutil==5.9.6

# Utilities
python-multipart==0.0.6
email-validator==2.1.0
python-dotenv==1.0.0
structlog==23.2.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
faker==20.1.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
pre-commit==3.6.0

# Monitoring
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
